Update image viewer to Responsive / User friendly OSS toolkit
This commit changes the screenshot gallery script to use the
open-source smartphoto.js instead.  This was chosen because
smartphoto offers a better user experience:

1) Smartphoto works excellently on mobile, the old solution
   needed a significant amount of work
2) Smartphoto plays nice with the browser (the back button
   will work as you would expect for example)
3) Smartphoto has better fallbacks for non-js users (links
   natively to the photo)
4) Smartphoto is MIT licensed so we can use it easily
5) Smartphoto is quite small when minified and will not incur
   a significant performance hit.

All in all, it would take a significant amount of time
to replicate Smartphoto's functionality with no upside to
doing so.  So I chose the path of least resistance.

Also added macOS .DS_Store files to .gitignore.